Refactor inventory sync script
The inventory sync script has been refactored to better accommodate object-oriented principles. A new class, InventorySync, has been defined to encapsulate the syncing functionality and allow for more modular and readable code. This allows for more maintainable code and better error handling.
